2025-06-22 - W171124MR/my_drf_project_1/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
PyMySQL==1.1.1
sqlparse==0.5.3
